Smarter handling of configuration file copies and restarts for postgres

+++ postgresql-9.5.sh
@@ -1,7 +1,9 @@
+restart=false
 if aptitude search '~i ^postgresql-9.5$' | grep -q postgresql-9.5; then
   echo "postgresql-9.5 already installed, skipping."
 else
   echo "installing postgres"
+  restart=true
   wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
   sh -c 'echo "deb http://apt.postgresql.org/pub/repos/apt/ precise-pgdg main" >> /etc/apt/sources.list.d/postgresql.list'
   apt-get update
@@ -9,15 +11,32 @@ else
   apt-get -y install postgresql-9.5 postgresql-server-dev-9.5 postgresql-contrib-9.5
 fi
 
-# Install configuration
-rm /etc/postgresql/9.5/main/postgresql.conf
-mv files/postgresql.conf /etc/postgresql/9.5/main/postgresql.conf
-chown postgres /etc/postgresql/9.5/main/postgresql.conf
+if cmp -s files/postgresql.conf /etc/postgresql/9.5/main/postgresql.conf; then
+  echo "postgresql.conf file is identical, skipping copy"
+else
+  restart=true
 
-# Enable client authentication
-rm /etc/postgresql/9.5/main/pg_hba.conf
-mv files/pg_hba.conf /etc/postgresql/9.5/main/pg_hba.conf
-chown postgres /etc/postgresql/9.5/main/pg_hba.conf
+  # Install configuration
+  rm /etc/postgresql/9.5/main/postgresql.conf
+  mv files/postgresql.conf /etc/postgresql/9.5/main/postgresql.conf
+  chown postgres /etc/postgresql/9.5/main/postgresql.conf
+fi
 
-/etc/init.d/postgresql restart
+if cmp -s files/pg_hba.conf /etc/postgresql/9.5/main/pg_hba.conf; then
+  echo "postgresql.conf file is identical, skipping copy"
+else
+  restart=true
+
+  # Enable client authentication
+  rm /etc/postgresql/9.5/main/pg_hba.conf
+  mv files/pg_hba.conf /etc/postgresql/9.5/main/pg_hba.conf
+  chown postgres /etc/postgresql/9.5/main/pg_hba.conf
+fi
+
+if [ $restart = true ]; then
+  /etc/init.d/postgresql restart
+else
+  echo "No update, skipping postgres restart"
+fi
+
